refactor(transaction): moved deployment infos to another file
